requirements.txt
pylast
yandex_music
pyyaml
PyQt5